New filter util service with a getAllSchedules with filters
