+++ src/main/java/io/boomerang/mongo/service/FlowWorkflowScheduleService.java
@@ -1,6 +1,7 @@
 package io.boomerang.mongo.service;
 
 import java.util.List;
+import java.util.Optional;
 import io.boomerang.mongo.entity.WorkflowScheduleEntity;
 import io.boomerang.mongo.model.WorkflowScheduleStatus;
 
@@ -18,6 +19,9 @@ public interface FlowWorkflowScheduleService {
       WorkflowScheduleStatus status);
 
   List<WorkflowScheduleEntity> getSchedules(List<String> ids);
+
+  List<WorkflowScheduleEntity> getAllSchedules(List<String> workflowIds,
+      Optional<List<String>> statuses, Optional<List<String>> types);
 
 
 }
