+++ src/main/java/io/boomerang/mongo/service/FlowWorkflowScheduleServiceImpl.java
@@ -1,7 +1,12 @@
 package io.boomerang.mongo.service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Service;
 import io.boomerang.mongo.entity.WorkflowScheduleEntity;
 import io.boomerang.mongo.model.WorkflowScheduleStatus;
@@ -12,6 +17,9 @@ public class FlowWorkflowScheduleServiceImpl implements FlowWorkflowScheduleServ
 
   @Autowired
   private FlowWorkflowScheduleRepository workflowScheduleRepository;
+  
+  @Autowired
+  private MongoTemplate mongoTemplate;
 
   @Override
   public void deleteSchedule(String id) {
@@ -41,5 +49,34 @@ public class FlowWorkflowScheduleServiceImpl implements FlowWorkflowScheduleServ
   @Override
   public List<WorkflowScheduleEntity> getSchedulesForWorkflowWithStatus(String workflowId, WorkflowScheduleStatus status) {
     return workflowScheduleRepository.findByWorkflowIdAndStatus(workflowId, status);
+  }
+
+  @Override
+  public List<WorkflowScheduleEntity> getAllSchedules(List<String> workflowIds,
+      Optional<List<String>> statuses, Optional<List<String>> types) {
+    
+    Criteria criteria = buildCriteriaList(workflowIds, types, statuses);
+    Query query = new Query(criteria);
+    List<WorkflowScheduleEntity> list = this.mongoTemplate.find(query, WorkflowScheduleEntity.class);
+    return list;
+  }
+
+  private Criteria buildCriteriaList(List<String> workflowIds,
+      Optional<List<String>> statuses, Optional<List<String>> types) {
+    List<Criteria> criterias = new ArrayList<>();
+    
+    if (types.isPresent()) {
+      Criteria dynamicCriteria = Criteria.where("type").in(types.get());
+      criterias.add(dynamicCriteria);
+    }
+    
+    if (statuses.isPresent()) {
+      Criteria dynamicCriteria = Criteria.where("status").in(statuses.get());
+      criterias.add(dynamicCriteria);
+    }
+    
+    Criteria workflowIdsCriteria = Criteria.where("workflowId").in(workflowIds);
+    criterias.add(workflowIdsCriteria);
+    return new Criteria().andOperator(criterias.toArray(new Criteria[criterias.size()]));
   }
 }
