+++ src/main/java/io/boomerang/service/FilterService.java
@@ -1,0 +1,11 @@
+package io.boomerang.service;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface FilterService {
+
+  List<String> getFilteredWorkflowIds(Optional<List<String>> workflowIds,
+      Optional<List<String>> teamIds, Optional<List<String>> scopes);
+
+}
